+++ src/biz/bokhorst/xprivacy/XCamera.java
@@ -9,7 +9,7 @@ public class XCamera extends XHook {
 	private Methods mMethod;
 
 	private XCamera(Methods method, String restrictionName) {
-		super(restrictionName, method.name(), null);
+		super(restrictionName, method.name(), "Camera." + method.name());
 		mMethod = method;
 	}